wpr-go: bugfix for binding server to any address
Previously, the WPR-go server offered a --host option that let you
specify which IP address to bind the WPR servers to. However, this
option was ignored, and only localhost could ever be listened on.

This CL corrects this so that now, for example, 0.0.0.0 can be specified
to accept non-localhost connections.

The CL also fixes a bug in reporting an error when initialising the
servers.

Change-Id: I6cdc8b89d1a7d68e0fb33c86433e6692b32f3c68
Bug: catapult:#3958
Reviewed-on: https://chromium-review.googlesource.com/832393
Commit-Queue: Tom Bergan <tombergan@chromium.org>
Reviewed-by: Helen Li <xunjieli@chromium.org>
Reviewed-by: Tom Bergan <tombergan@chromium.org>